Move useAWS to public header
