+++ include/rocksdb/cloud/cloud_file_system.h
@@ -26,6 +26,13 @@ class S3Client;
 
 namespace ROCKSDB_NAMESPACE {
 
+#ifdef USE_AWS
+// Initializes AWS, or joins an existing initialization. Keep the
+// returned shared_ptr alive for as long as you will be using AWS
+// operations
+std::shared_ptr<void> useAWS();
+#endif
+
 class CloudFileSystem;
 class CloudLogController;
 class CloudStorageProvider;
